+++ src/Products/PloneMeeting/skins/plonemeeting_styles/plonemeetingpopups.js
@@ -381,7 +381,7 @@ function inheritedItemInfos() {
 function usersGroupInfos() {
     tooltipster_helper(selector='.tooltipster-group-users',
                        view_name='@@display-group-users',
-                       data_parameters=['group_ids'],
+                       data_parameters=['group_ids:json'],
                        options={position: 'bottom'});
 }